requirements.txt
pinecone-client>=2.2.1
sentence-transformers>=2.2.2
unstructured>=0.10.0
pypdf>=3.15.0
python-dotenv>=1.0.0
Flask>=2.3.0
requests>=2.31.0
torch>=2.0.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
tqdm>=4.66.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
openpyxl>=3.1.2
pandas>=2.0.0
ebooklib>=1.0.0
python-pptx>=0.6.21
python-rtf>=0.4.5
lxml>=4.9.0 